requirements.txt
clickhouse-driver==0.2.0
numpy==1.19.4
psycopg2-binary==2.8.6
pytz==2020.5
tzlocal==2.1
